Move user seeder to user controller
