+++ controllers/user/index.js
@@ -1,6 +1,9 @@
 'use strict';
 const express = require('express');
+const uuidv4 = require('uuid/v4');
+const features = require('config').get('features');
 
+const { Users } = require('../../db/models');
 const { redirectForLocale } = require('../../common/urls');
 const { noCache } = require('../../middleware/cached');
 const { noindex } = require('../../middleware/robots');
@@ -10,8 +13,37 @@ const router = express.Router();
 
 router.use(noCache, noindex);
 
+/**
+ * Staff auth
+ * Azure Active Directory authentication for staff members
+ */
 router.use('/staff', require('./staff'));
 
+/**
+ * User seed endpoint
+ * Allows generation of seed users in test environments
+ * Seed users have isActive automatically set as true
+ * to bypass activation flow in tests.
+ */
+if (features.enableSeeders) {
+    router.post('/seed', (req, res) => {
+        const username = `${uuidv4()}@example.com`;
+        const password = uuidv4();
+
+        Users.createUser({
+            username: username,
+            password: password,
+            isActive: true
+        }).then(() => {
+            res.json({ username, password });
+        });
+    });
+}
+
+/**
+ * Public user routes
+ * Disallow staff from this point on
+ */
 router.use(requireNotStaffAuth, function(req, res, next) {
     res.locals.bodyClass = 'has-static-header'; // No hero images on user pages
     res.locals.sectionTitle = req.i18n.__('user.common.yourAccount');
